Do not generate auto requires for rubygem-nokogiri
With auto requires created, a dependency on rubygem-racc is generated.
The gem racc is bundled into ruby-libs and is provided via
bundled(rubygem-racc). This will lead to a failure to install
as rubygem-racc cannot be resolved on install of this package
on EL8.

(cherry picked from commit ee3fc2d36886c61848ff744351bf1a2833f40c76)